# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stores/{id}/items": {
      "get": {
        "operationId": "StoreItemController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedStoreItemResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List store items with pagination",
        "tags": [
          "Store Item"
        ]
      }
    },
...
    "/families": {
      "post": {
        "operationId": "FamilyController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFamilyRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new family",
        "tags": [
          "Family"
        ]
      },
      "get": {
        "operationId": "FamilyController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all families for the current user",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/{id}": {
      "get": {
        "operationId": "FamilyController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyWithMembersResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a family with its members",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/{id}/invitations": {
      "post": {
        "operationId": "FamilyController_inviteMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyInvitationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite a member to a family",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/invitations/pending": {
      "get": {
        "operationId": "FamilyController_getPendingInvitations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyInvitationResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get pending invitations for the current user",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/{id}/invitations/pending": {
      "get": {
        "operationId": "FamilyController_getFamilyPendingInvitations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyInvitationResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get pending invitations for a family",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/invitations/{id}/accept": {
      "post": {
        "operationId": "FamilyController_acceptInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyMemberResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept a family invitation",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/invitations/{id}/decline": {
      "post": {
        "operationId": "FamilyController_declineInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Decline a family invitation",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/invitations/{id}/cancel": {
      "post": {
        "operationId": "FamilyController_cancelInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a family invitation",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/{id}/members/{userId}": {
      "delete": {
        "operationId": "FamilyController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a member from a family",
        "tags": [
          "Family"
        ]
      }
    },
    "/families/{id}/members/me": {
      "delete": {
        "operationId": "FamilyController_leaveFamily",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Leave a family",
        "tags": [
          "Family"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fynans API",
    "description": "Fynans personal finance API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StoreItemSizeDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          }
        },
        "required": [
          "value",
          "unit"
        ]
      },
      "StoreItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "isDiscounted": {
            "type": "boolean"
          },
          "categoryId": {
            "type": "string"
          },
          "size": {
            "$ref": "#/components/schemas/StoreItemSizeDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "storeId",
          "name",
          "price",
          "isDiscounted",
          "categoryId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedStoreItemResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoreItemResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
...
      "CreateFamilyRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ]
      },
      "FamilyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "balance",
          "createdAt",
          "updatedAt"
        ]
      },
      "FamilyMemberUserInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName"
        ]
      },
      "FamilyMemberUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "role": {
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER"
            ],
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/FamilyMemberUserInfoDto"
          }
        },
        "required": [
          "id",
          "userId",
          "role",
          "balance",
          "joinedAt",
          "user"
        ]
      },
      "FamilyWithMembersResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyMemberUserDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "balance",
          "createdAt",
          "updatedAt",
          "members"
        ]
      },
      "InviteMemberRequestDto": {
        "type": "object",
        "properties": {
          "inviteeEmail": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "inviteeEmail"
        ]
      },
      "FamilyInvitationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "familyId": {
            "type": "string"
          },
          "inviterId": {
            "type": "string"
          },
          "inviteeId": {
            "type": "string"
          },
          "inviteeEmail": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED",
              "CANCELLED"
            ]
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "familyId",
          "inviterId",
          "inviteeEmail",
          "status",
          "expiresAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "FamilyMemberResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "familyId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER"
            ]
          },
          "balance": {
            "type": "number"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "familyId",
          "userId",
          "role",
          "balance",
          "joinedAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}